Move literalInput.ts and customExprSelector.ts to components subfolder
